+++ akash/client/cli/groupings.go
@@ -1,7 +1,7 @@
 package cli
 
 func (c AkashCommand) DefaultGas() AkashCommand {
-	return c.GasAuto().SetGasAdjustment(1.15).SetGasPrices().SetSignMode("amino-json")
+	return c.GasAuto().SetGasAdjustment(1.5).SetGasPrices().SetSignMode("amino-json")
 }
 
 func (c AkashCommand) SetSeqs(dseq string, gseq string, oseq string) AkashCommand {