+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Bug where destroying deployments did not really close them
+### Development
+- Replaced HTTP API call for deployments with CLI
 
 ## [0.0.4] - 2022-08-17
 ### Added