+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.6] - 2022-11-16
+### Added
+- Add support to mainnet-4
+
+## [0.0.5] - 2022-10-27
 ### Added
 - Transaction memo stating the Terraform provider was used and which version
 ### Fixed